changing deployment script name
